fixed most things

+++ src/fast_hmm_param.h
@@ -28,11 +28,11 @@ struct fast_hmm_param{
         float** transition;
         float** emission;
         float* background_emission;
-        uint16_t last_state;
-        uint32_t alloc_items;
-        uint32_t num_items;
-        uint32_t alloc_num_states;
-        uint32_t L;
+        int last_state;
+        int alloc_items;
+        int num_items;
+        int alloc_num_states;
+        int L;
 };
 
 /* For when we want to train multiple models  */
